+++ src/Standard/Standard_UUID.hxx
@@ -18,18 +18,17 @@
 #include <Standard_Type.hxx>
 
 #ifdef _WIN32
-#include <windows.h>
+  #include <windows.h>
 #else
-typedef struct {
-  unsigned long Data1 ;
-  unsigned short Data2 ;
-  unsigned short Data3 ;
-  unsigned char Data4[8] ;
-} GUID ;
+  typedef struct
+  {
+    unsigned int   Data1;
+    unsigned short Data2;
+    unsigned short Data3;
+    unsigned char  Data4[8];
+  } GUID;
 #endif
 
-typedef GUID Standard_UUID ;
+typedef GUID Standard_UUID;
 
 #endif
-
-
